Add NaviPlugin (a collection of listeners)
While adding support for navi for a library I struggled creating a simple way to bind to all events I need. I ended up creating my library object and binding it multiple times with `addListener(myLib);` to the NaviActivity. In the README for my lib I have to add "add those 10 lines to make my lib work". This adds to much complexity and show implementation details to the implementer which is not needed.
Even worse: When updating my lib, it may happen that I have to add a new listener. This would require the user of my lib to update the version AND update the listeners where my lib is used.

With this plugin feature it's more like "simply add `addPlugin(new MyPlugin());` in you NaviActivity and you are fine".

ActivityNaviPlugin is the default implementation. A plugin for Fragment will follow